requirements.txt
django-bootstrap4
django-extensions
djangorestframework
